refactor: add WhisperBanner
